% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpmodel-class.R
\docType{methods}
\name{bcgp_sampling,bcgpmodel-method}
\alias{bcgp_sampling,bcgpmodel-method}
\title{Draw samples from a BCGP model}
\usage{
\S4method{bcgp_sampling}{bcgpmodel}(object, algorithm = c("NUTS", "MH"),
  scaled = TRUE, chains = 4L, init = "random", numUpdates = 5,
  nAdapt = 1000, burnin = 1000, nmcmc = 10000, ...)
}
\arguments{
\item{object}{An instance of S4 class \code{bcgpmodel}.}

\item{algorithm}{Either \code{"NUTS"} for the \emph{No U-Turn Sampler}
implemented by Stan, or \code{"MH"} for a Metropolis-Hastings algorithm}

\item{scaled}{A logical indicating whether the data should be scaled before
fitting. It is highly recommended to scale the data before fitting.}

\item{chains}{A positive integer specifying the number of Markov chains}

\item{init}{Can be either the string "random" or a list of length
\code{chains}. The elements of this list should be named lists, where each of
these has the name of a parameter and is used to specify the initial values
for that parameter for the corresponding chain.

\describe{
\code{init = "random"} (default):
The initial values will be generated randomly from their respective prior
distributions.

\code{init} via list:
Set initial values by providing a list equal in length to the number of
Markov chains. A call to \code{create_Inits()} will assist in the correct
creation of this list.
}}

\item{numUpdates}{A positive integer for the number of updates in the
proposal stepsize adaptation phase. Ignored if \code{algorithm = "NUTS"}.}

\item{burnin}{A positive integer for the number of burnin samples to discard
after the stepsize adaptation phase is finished. This is equivalent to the
parameter \code{warmup} in \code{\link[rstan]{stan}}}

\item{nmcmc}{The number of samples to be kept for each Markov chain.}

\item{...}{optional parameters, only if \code{algorithm = "NUTS"}. See the
documentation for \code{\link[rstan]{stan}}. Arguments not listed here will
be ignored.
\describe{
  \item{\code{control}}{A named list of parameters to control the sampler's
  behavior. It defaults to NULL so all the default values are used. }
  \item{\code{gridTest}}{A logical indicating whether the test data should be
  generated on a grid. Defaults to \code{FALSE}. Only useful for
  \eqn{d \geq 2}.}
  \item{\code{gridTestSize}}{An integer indicating the number of points per
  dimension for the test grid. Only useful for \eqn{d \geq 2}.}
  \item{}{Be aware that a grid in high dimensions quickly gets very large.
  For example, for \code{d} = 3 and \code{gridTestSize} = 10, the number of
  points in this grid is \eqn{10^3 = 1000}. Therefore, in higher dimensions (
  \eqn{d > 4}), the simulation will default to \code{gridTest = FALSE} if
  \code{gridTestSize} is left unspecified, and \code{nTest} test locations
  will be randomly selected on \eqn{[0, 1]^d}.}
}}

\item{numAdapt}{A positive integer for the number of samples within each
update in the proposal stepsize adaptation phase. Ignored if
\code{algorithm = "NUTS"}.}
}
\value{
An instance of S4 class \code{bcgpfit}. It contains information
about the data, model, sampling algorithm, and sample draws from
the posterior.
}
\description{
\code{bcgp_sampling} draws samples from the model defined by class
\code{bcgpmodel}
}
\details{
This returns an instance of S4 class \code{bcgpfit}. It contains information
about the data, model, sampling algorithm, and sample draws from
the posterior
}
