% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpmodel-class.R
\name{bcgp_sampling}
\alias{bcgp_sampling}
\title{setGeneric(name = "update_inits",
           def = function(object, ...) standardGeneric("update_inits"))}
\usage{
bcgp_sampling(object, ...)
}
\description{
#' Update the starting values for MCMC chains
#'
#' \code{update_inits} returns an instance of S4 class \code{bcgpmodel} after
#' updating the starting values for each Markov chain
#'
#' This returns an instance of S4 class \code{bcgpmodel}. It is generally meant
#' to update the starting values for the Markov chains after the user has
#' changed the prior distributions in an instance of class \code{bcgpmodel}. The
#' new initial values are simulated from the prior distribution, ensuring they
#' are within the support of the posterior distribution.
#'
#' @param object An instance of S4 class \code{bcgpmodel}.
#' @return An instance of S4 class \code{bcgpmodel} containing initial values
#' randomly from the prior distributions for all the parameters
#' @family preprocessing functions
#' @seealso \linkS4class{bcgpmodel}
#' @examples
#' simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
#' model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
#'                    composite = TRUE, stationary = FALSE, noise = FALSE,
#'                    scaled = TRUE, chains = 4L)
#' model@priors$w$lower <- 0.75
#' model <- update_inits(model)
#' @export
}
\details{
setMethod("update_inits", signature = "bcgpmodel",
          function(object){

            data <- object@data

            if(object@scaled){
              xIn <- data$scaled$x
            }else{
              xIn <- data$raw$x
            }

            object@inits <- create_inits(x = xIn, object@composite,
                                         object@stationary, object@noise,
                                         object@priors, object@chains)


            return(object)

          })
}
