% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcgpfit-class}
\alias{bcgpfit-class}
\title{An S4 class to represent a fitted BCGP model}
\description{
This class contains the data, both raw and scaled, information about the
distributions for the parameters, the values of the hyperparameters, the
number of chains desired to be fit, whether the data was scaled to
\eqn{[0, 1]^d} before fitting,
along with some other information about the type of model fit
(composite/non-composite, stationary/non-stationary, deterministic/noisy).
}
\section{Slots}{

\describe{
\item{\code{model_name}}{A character describing the type of model that was fit.
Corresponds to the values for \code{stationary} and \code{composite}.}

\item{\code{data}}{A list that contains the training data. One element of the list
contains the raw data, and one element contains the scaled data in which the
independent variables are scaled to \eqn{[0, 1]^d}, and the response variable
is scaled to have mean 0 and variance 1.}

\item{\code{composite}}{A logical indicating whether the model is composite or not.}

\item{\code{stationary}}{A logical indicating whether the model is stationary or not.}

\item{\code{noise}}{A logical indicating whether the data is noisy or deterministic
(as from a computer model).}

\item{\code{scaled}}{A logical indicating whether the data was scaled before fitting.
It is highly recommended to scale the data before fitting.}

\item{\code{chains}}{A positive integer specifying the number of Markov chains}

\item{\code{priors}}{A list with an element for each parameter in the BCGP model
specified by \code{composite} and \code{stationary}. Each element contains
the values of the hyperparameters for each parameter.}

\item{\code{distributions}}{A list with an element for each parameter in the BCGP
model specified by \code{composite} and \code{stationary}. Each element
contains a character string identifying the prior distribution for each
parameter.}

\item{\code{init}}{A list of length \code{chains} that contains initial values for
each parameter for the MCMC algorithm.}

\item{\code{model_pars}}{A character vector giving the parameter names}

\item{\code{par_dims}}{A list giving the dimension of each parameter}

\item{\code{sim}}{A list containing the posterior samples, along with other
information}

\item{\code{algorithm}}{Either \code{"NUTS"} for the \emph{No U-Turn Sampler}
implemented by Stan, or \code{"MH"} for a Metropolis-Hastings algorithm}

\item{\code{sampler_args}}{A list containing information about the sampling
algorithm.}
}}

\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
model <- bcgpmodel(x = simData@training$x, y = simData@training$y
                   composite = TRUE, stationary = FALSE, noise = TRUE)
\dontrun{
bcgp_sampling(model, algorithm = "NUTS", scaled = TRUE, chains = 3L,
              cores = 1L, control = list(adapt_delta = 0.99,
                                         max_treedepth = 15))
}
}
\seealso{
\code{\link{bcgp_sampling}} \linkS4class{bcgpmodel}
\code{\link{bcgpmodel}}
}
