% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcgpmodel-class}
\alias{bcgpmodel-class}
\title{An S4 class to represent a BCGP model}
\description{
This class contains the data, information about the distributions for the
parameters, the values of the hyperparameters, along with some other
information about the type of model desired to be fit
(composite/non-composite, stationary/non-stationary, deterministic/noisy).
}
\section{Slots}{

\describe{
\item{\code{data}}{A list that contains the training data. One element of the list
contains the raw data, and one element contains the scaled data in which the
independent variables are scaled to \eqn{[0, 1]^d}, and the response variable
is scaled to have mean 0 and variance 1.}

\item{\code{composite}}{A logical indicating whether the model is composite or not.}

\item{\code{stationary}}{A logical indicating whether the model is stationary or not.}

\item{\code{noise}}{A logical indicating whether the data is noisy or deterministic
(as from a computer model).}

\item{\code{priors}}{A list with an element for each parameter in the BCGP model
specified by \code{composite} and \code{stationary}. Each element contains
the values of the hyperparameters for each parameter.}

\item{\code{distributions}}{A list with an element for each parameter in the BCGP
model specified by \code{composite} and \code{stationary}. Each element
contains a character string identifying the prior distribution for each
parameter.}
}}

\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
bcgpmodel(x = simData@training$x, y = simData@training$y
          composite = TRUE, stationary = FALSE, noise = TRUE)
}
\seealso{
\code{\link{bcgpmodel}}
}
