% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpmodel-class.R
\name{bcgpmodel}
\alias{bcgpmodel}
\title{Create an object of class bcgpmodel}
\usage{
bcgpmodel(x, y, composite = TRUE, stationary = FALSE, noise = FALSE)
}
\arguments{
\item{x}{An \code{n x d} numeric matrix, where \code{d} is the dimension of
the data, and \code{n} is the number of training data observations.}

\item{y}{A numeric vector of length \code{n}.}

\item{composite}{A logical, \code{TRUE} for a composite of a global process,
a local process, and an error process, \code{FALSE} for non-composite.
Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data is known to be noise-free (such as from a
deterministic computer model), then \code{noise} should be \code{FALSE}.
Otherwise, it should be \code{TRUE}. Defaults to \code{FALSE}}
}
\value{
An instance of S4 class \code{bcgpmodel} containing the data, default
values for all the prior parameters, information about the process, and
information about the distributions.
}
\description{
\code{bcgpmodel} returns an instance of S4 class \code{bcgpmodel}
}
\details{
This creates an instance of S4 class \code{bcgpmodel} that contains
user-provided data, user-provided information about the type of model desired
to be fit (composite/non-composite, stationary/non-stationary,
deterministic/noisy), default values for the hyperparameters for the prior
distributions, and information about the prior distributions.
}
\examples{
data_sim <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
bcgpmodel(x = data_sim@training$x, y = simData@training$y,
          composite = TRUE, stationary = FALSE, noise = FALSE)
}
\seealso{
\linkS4class{bcgpmodel} \code{\link{bcgpfit}}
\code{\link{bcgp}}

Other preprocessing functions: 
\code{\link{create_parameter_list}()}
}
\concept{preprocessing functions}
