% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcgppriors-class}
\alias{bcgppriors-class}
\title{An S4 class to represent priors for a BCGP model}
\description{
This class contains information about the distributions for the parameters
and the values of the hyperparameters.
}
\section{Slots}{

\describe{
\item{\code{priors}}{A list with an element for each parameter in the BCGP model
specified by \code{composite} and \code{stationary}. Each element contains
the values of the hyperparameters for each parameter.}

\item{\code{distributions}}{A list with an element for each parameter in the BCGP
model specified by \code{composite} and \code{stationary}. Each element
contains a character string identifying the prior distribution for each
parameter.}

\item{\code{stationary}}{A logical indicating whether the model is stationary or not.}

\item{\code{composite}}{A logical indicating whether the model is composite or not.}

\item{\code{noise}}{A logical indicating whether the data is noisy or deterministic
(as from a computer model).}
}}

\examples{
create_priors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
bcgppriors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
}
\seealso{
\code{\link{create_priors}} \code{\link{bcgppriors}}
}
