% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpinits-class.R, R/bcgppriors-class.R
\name{bcgppriors}
\alias{bcgppriors}
\title{Create an object of class bcgpinits}
\usage{
bcgppriors(composite = TRUE, stationary = FALSE, noise = FALSE,
  d = 1L)

bcgppriors(composite = TRUE, stationary = FALSE, noise = FALSE,
  d = 1L)
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process,
a local process, and an error process, \code{FALSE} for non-composite.
Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}

\item{object}{An instance of S4 class \code{bcgppriors}. Either the function
\code{bcgppriors()} or \code{create_priors()} should be called to create this
object, which can then be modified before inputting into \code{bcgpinits()}.}

\item{chains}{An integer specifiying the number of chains desired in the MCMC
algorithm.}
}
\value{
An instance of S4 class \code{bcgpinits} containing randomly
generated initial values for all the parameters, information about the prior
distributions, and information about the process.

An instance of S4 class \code{bcgppriors} containing the default
values for all the prior parameters, information about the process, and
information about the distributions.
}
\description{
\code{bcgpinits} returns an instance of S4 class \code{bcgpinits}

\code{bcgppriors} returns an instance of S4 class \code{bcgppriors}
}
\details{
This creates an instance of S4 class \code{bcgpinits}. The list (of length
\code{chains}) of initial values is simulated from the prior distributions
for each parameter. The returned object also contains information for the
priors, information about the process, and information about the prior
distributions. The user can change the values as they like prior to
fitting any data with \code{\link{bcgp}}. \code{create_inits()} can also be
called to create a \code{bcgpinits} object.

This creates an instance of S4 class \code{bcgppriors} that contains default
values for the priors, information about the process, and information about
the distributions. The user can change the values as they like prior to
fitting any data with \code{\link{bcgp}}. \code{create_priors()} can also be
called to create a \code{bcgppriors} object.
}
\examples{
bcgpinits(object = )
bcgppriors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
bcgppriors(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1)
bcgppriors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
}
\seealso{
\linkS4class{bcgpinits} \code{\link{create_inits}}
\linkS4class{bcgppriors} \code{\link{bcgp}}

\linkS4class{bcgppriors} \code{\link{create_priors}}
\code{\link{bcgp}}

Other preprocessing functions: \code{\link{createInits}},
  \code{\link{create_inits,bcgppriors-method}},
  \code{\link{create_parameter_list}},
  \code{\link{create_priors}}

Other preprocessing functions: \code{\link{createInits}},
  \code{\link{create_inits,bcgppriors-method}},
  \code{\link{create_parameter_list}},
  \code{\link{create_priors}}
}
\concept{preprocessing functions}
