% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BJX}
\alias{BJX}
\title{Data from a non-stationary function}
\format{A list with components for training data and test data:
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{x_test}{test data locations}
  \item{y_test}{test data values}
}}
\usage{
BJX
}
\description{
A dataset containing data simulated from a non-stationary function used in
Xiong (2007) and Ba and Joseph (2009)
}
\keyword{datasets}
