% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cox_mats_functions.R
\name{combine_cor_mats_R}
\alias{combine_cor_mats_R}
\title{Combine correlation matrices.}
\usage{
combine_cor_mats_R(w, G, L)
}
\arguments{
\item{w}{A scalar between 0 and 1. In the \code{bcgp} setting, the user will
have set a lower and upper bound (the default is to be between 0.5 and 1).}

\item{G}{An \emph{n x n} correlation matrix, often the result of
\code{\link{get_cor_mat_R}}.}

\item{L}{An \emph{n x n} correlation matrix, often the result of
\code{\link{get_cor_mat_R}}.}
}
\value{
An \emph{n x n} correlation matrix
}
\description{
\code{combine_cor_mats_R} returns a correlation matrix that is a weighted sum
 of two other correlation matrices.
}
\details{
This creates a correlation matrix, \emph{R}, where \deqn{R = wG + (1-w)L}. In
the \code{bcgp} setting, \emph{G} is the global correlation matrix, \emph{L}
is the local correlation matrix, and \emph{w} is the weight.
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rhoG <- runif(d, 0, 1)
rhoL <- runif(d, 0, rhoG)
G <- get_cor_mat_R(x, rhoG)
L <- get_cor_mat_R(x, rhoL)
w <- runif(1, 0.5, 1)
combine_cor_mats_R(w, G, L)
}
\seealso{
Other correlation and covariance functions: 
\code{\link{get_cor_mat_R}()},
\code{\link{get_cov_mat_ns_R}()},
\code{\link{get_cov_mat_s_R}()}
}
\concept{correlation and covariance functions}
