% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxMatsR.R
\name{combineCorMatsR}
\alias{combineCorMatsR}
\title{Combine correlation matrices.}
\usage{
combineCorMatsR(w, G, L)
}
\arguments{
\item{w}{A scalar between 0 and 1. In the \code{bcgp} setting, the user will
have set a lower and upper bound (the default is to be between 0.5 and 1).}

\item{G}{An \emph{n x n} correlation matrix, often the result of
\code{\link{getCorMat}}.}

\item{L}{An \emph{n x n} correlation matrix, often the result of
\code{\link{getCorMat}}.}
}
\value{
An \emph{n x n} correlation matrix
}
\description{
\code{combineCorMatsR} returns a correlation matrix that is a weighted sum of
two other correlation matrices.
}
\details{
This creates a correlation matrix, \emph{R}, where \deqn{R = wG + (1-w)L}. In
the \code{bcgp} setting, \emph{G} is the global correlation matrix, \emph{L}
is the local correlation matrix, and \emph{w} is the weight.
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rhoG <- runif(d, 0, 1)
rhoL <- runif(d, 0, rhoG)
G <- getCorMat(x, rhoG)
L <- getCorMat(x, rhoL)
w <- runif(1, 0.5, 1)
combineCorMatsR(w, G, L)
}
\seealso{
Other correlation and covariance functions: \code{\link{getCorMatR}},
  \code{\link{getCovMatNSR}}, \code{\link{getCovMatSR}}
}
\concept{correlation and covariance functions}
