% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgppriors-class.R
\docType{methods}
\name{create_inits,bcgppriors-method}
\alias{create_inits,bcgppriors-method}
\title{Create an object of class bcgpinits}
\usage{
\S4method{create_inits}{bcgppriors}(object, chains = 4L, ...)
}
\arguments{
\item{object}{An instance of S4 class \code{bcgppriors}. Either the function
\code{bcgppriors()} or \code{create_priors()} should be called to create this
object, which can then be modified before inputting into
\code{create_inits()}.}

\item{chains}{An integer specifiying the number of chains desired in the MCMC
algorithm.}

\item{...}{optional parameters
\describe{
  \item{\code{x}}{An \code{n x d} matrix of training data locations. This
  matrix is required for a nonstationary model in order to simulate the
  variance process at the training data locations.}
  }}
}
\value{
An instance of S4 class \code{bcgpinits} containing randomly
generated initial values for all the parameters, information about the prior
distributions, and information about the process.
}
\description{
\code{create_inits} returns an instance of S4 class \code{bcgpinits}
}
\details{
This creates an instance of S4 class \code{bcgpinits}. The list (of length
\code{chains}) of initial values is simulated from the prior distributions
for each parameter. The returned object also contains information for the
priors, information about the process, and information about the prior
distributions. The user can change the values as they like prior to
fitting any data with \code{\link{bcgp}}. \code{bcgpinits()} can also be
called to create a \code{bcgpinits} object.
}
\examples{
create_inits(object = create_priors(), chains = 4,
             x = matrix(seq(0, 1, length.out = 15), ncol = 1))
create_inits(object = create_priors(stationary = TRUE), chains = 4)
}
\seealso{
\linkS4class{bcgpinits} \code{\link{bcgpinits}}
\linkS4class{bcgppriors} \code{\link{bcgp}}

Other preprocessing functions: \code{\link{bcgpinits}},
  \code{\link{bcgppriors}}, \code{\link{create_inits_raw}},
  \code{\link{create_parameter_list}},
  \code{\link{create_priors}}
}
\concept{preprocessing functions}
