% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFuncs.R
\name{create_inits}
\alias{create_inits}
\title{Create a list with initial values.}
\usage{
create_inits(x, composite, stationary, noise, priors, chains)
}
\arguments{
\item{x}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{priors}{A list that contains the parameter values for the priors.}

\item{chains}{The number of Markov chains. The default is 4.}
}
\value{
A list of length \code{chains} The elements of this list will be named
lists, where each of these has the name of a parameter.
}
\description{
\code{create_inits} returns a list that contains initial values randomly
generated from the prior distribution.
}
\details{
This creates a list of length \code{chains} that contains randomly generated
initial values. The intention is to specify initial values for the parameters
for the Markov chains. The user can change the values as they like prior to
inputting the list into \code{bcgp}.
}
\section{TODO}{
 Decide whether to add options for "heteroscedastic" and "composite"
}

\examples{
x <- matrix(runif(40), ncol= 4, nrow = 10)
create_inits(x)
create_inits(x, priors = create_priors(), chains = 2)
}
\seealso{
\code{\link{bcgp}}
}
