% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_priors_functions.R
\name{create_priors}
\alias{create_priors}
\title{Create a list of priors for BCGP}
\usage{
create_priors(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1L)
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process,
a local process, and an error process, \code{FALSE} for non-composite.
Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}
}
\value{
A list containing the default values for all the prior parameters,
information about the process, and information about the prior distributions.
}
\description{
\code{create_priors} returns a list of default priors.
}
\details{
This creates a list that contains default hyperparameter values for a
model corresponding to inputs \code{composite}, \code{stationary},
\code{noise}, and \code{d}. This function is meant to set up the priors in a
\code{bcgpmodel} object before a call to \code{bcgp_sampling()}. The
user will be able to modify the hyperparameter values in the \code{bcgpmodel}
object before the call to \code{bcgp_sampling()} if desired, but the user
has no choice of prior distributions (only the parameters that define the
given distribution).
}
\examples{
create_priors(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1)
create_priors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
}
\seealso{
\linkS4class{bcgpmodel} \code{\link{bcgpfit}}
\code{\link{bcgp}}
}
