% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFuncs.R
\name{createParameterList}
\alias{createParameterList}
\title{Create a list with random parameter values.}
\usage{
createParameterList(composite = TRUE, stationary = FALSE,
  noise = FALSE, d = 1, ...)
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process
and a local process, \code{FALSE} for non-composite. Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}

\item{...}{Arguments \code{n}, \code{nPred}, \code{x}, and \code{xPred} that
are only necessary for non-stationary processes. If left unspecified, lack of
input for \code{n} and \code{nPred} will throw an error. If \code{x} and
\code{xPred} are left unspecified, these matrices will be randomly generated
on \eqn{[0, 1]^d}. If specified, \code{x} will be scaled to \eqn{[0, 1]^d}
before variance parameters are generated, and \code{xPred} will be scaled
based on the scaling of {x} before variance parameters are generated.}
}
\description{
\code{createParameterList} returns a list that contains randomly generated
parameter values.
}
\details{
This creates a list that contains randomly generated parameter values for a
model corresponding to inputs \code{composite}, \code{stationary},
\code{noise}, and \code{d}. This function is meant to set up the parameters
before a call to \code{simulate_from_model}. The user can modify the list of
parameter values if desired.
}
\examples{
createParameterList(composite = FALSE, stationary = TRUE, noise = FALSE, d = 1)
createParameterList(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1,
                    n = 15, nPred = 100)
}
\seealso{
\code{\link{simulate_from_model}}

Other preprocessing functions: \code{\link{createInits}},
  \code{\link{createPriors}}
}
\concept{preprocessing functions}
