% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cox_mats_functions.R
\name{get_cor_mat_R}
\alias{get_cor_mat_R}
\title{Create a correlation matrix.}
\usage{
get_cor_mat_R(x, rho)
}
\arguments{
\item{x}{An \code{n x d} matrix, where \code{d} is the dimension of the data.}

\item{rho}{A vector of length d. Each value in \code{rho} should be between
0 and 1.}
}
\value{
An \code{n x n} correlation matrix
}
\description{
\code{get_cor_mat_R} returns a correlation matrix.
}
\details{
This creates a correlation matrix, \code{R}, where \deqn{R_{ij} = \prod_{k =
1}^{d}\rho_k^{16\left( x_{ik} - x_{jk} \right)^2}}
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
get_cor_mat_R(x, rho)
}
\seealso{
Other correlation and covariance functions: 
\code{\link{combine_cor_mats_R}()},
\code{\link{get_cov_mat_ns_R}()},
\code{\link{get_cov_mat_s_R}()}
}
\concept{correlation and covariance functions}
