% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cox_mats_functions.R
\name{get_cov_mat_s_R}
\alias{get_cov_mat_s_R}
\title{Create a covariance matrix.}
\usage{
get_cov_mat_s_R(sigma2, R, sigma2eps)
}
\arguments{
\item{sigma2}{A positive scalar representing the process variance}

\item{R}{An \emph{n x n} correlation matrix.}

\item{sigma2eps}{A positive scalar representing the variance of the noise
(or nugget).}
}
\value{
An \code{n x n} covariance matrix
}
\description{
\code{get_cov_mat_s_R} returns a stationary covariance matrix (with nugget).
This creates a covariance matrix, C, where \deqn{C =  sigma^2 R + \sig2Eps I}
\code{R} is a correlation matrix, \code{sigma2} the process variance, and
\code{sigma2eps} is the variance of the noise (or nugget).
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
R <- get_cor_mat_R(x, rho)
sigma2 <- 1.2
get_cov_mat_s_R(sigma2, R, sigma2eps)
}
\seealso{
Other correlation and covariance functions: 
\code{\link{combine_cor_mats_R}()},
\code{\link{get_cor_mat_R}()},
\code{\link{get_cov_mat_ns_R}()}
}
\concept{correlation and covariance functions}
