% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxMatsR.R
\name{getCorMatR}
\alias{getCorMatR}
\title{Create a correlation matrix.}
\usage{
getCorMatR(x, rho)
}
\arguments{
\item{x}{An \emph{n x d} matrix, where \emph{d} is the dimension of the data.}

\item{rho}{A vector of length d. Each value in \code{rho} should be between
0 and 1.}
}
\value{
An \emph{n x n} correlation matrix
}
\description{
\code{getCorMatR} returns a correlation matrix.
}
\details{
This creates a correlation matrix, \emph{R}, where \deqn{R_{ij} = \prod_{k =
1}^{d}\rho_k^{16\left( x_{ik} - x_{jk} \right)^2}}{R_ij = \prod_{k =
1}^{d}\rho_k^{16\left( x_{ik} - x_{jk} \right)^2}}
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
getCorMatR(x, rho)
}
\seealso{
Other correlation and covariance functions: \code{\link{combineCorMatsR}},
  \code{\link{getCovMatNSR}}, \code{\link{getCovMatSR}}
}
\concept{correlation and covariance functions}
