% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxMatsR.R
\name{getCovMatSR}
\alias{getCovMatSR}
\title{Create a covariance matrix.}
\usage{
getCovMatSR(sigma2, R, sigma2eps)
}
\arguments{
\item{sigma2}{A positive scalar representing the process variance}

\item{R}{An \emph{n x n} correlation matrix.}

\item{sig2eps}{A positive scalar representing the variance of the noise
(or nugget).}
}
\value{
An \emph{n x n} covariance matrix
}
\description{
\code{getCovMatSR} returns a stationary covariance matrix (with nugget).
This creates a covariance matrix, C, where \deqn{C =  sigma^2 R + \sig2Eps I}
\emph{R} is a correlation matrix, \emph{sigma2} the process variance, and
\emph{sig2eps} is the variance of the noise (or nugget).
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
R <- getCorMat(x, rho)
sigma2 <- 1.2
getCovMatSR(V, R, sig2)
}
\seealso{
Other correlation and covariance functions: \code{\link{combineCorMatsR}},
  \code{\link{getCorMatR}}, \code{\link{getCovMatNSR}}
}
\concept{correlation and covariance functions}
