% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HE}
\alias{HE}
\title{Heat Exchange Data}
\format{A list with elements \code{x, y, xPred, yPred}
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{xPred}{test data locations}
  \item{yPred}{test data values}
}}
\source{
TODO: add source
}
\usage{
HE
}
\description{
A dataset containing the training data and test data from
the heat exchange example used in the paper
}
\keyword{datasets}
