% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{HeatExchange}
\alias{HeatExchange}
\title{Data from the Heat Exchange example}
\format{A list with components for training data and test data:
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{x_test}{test data locations}
  \item{y_test}{test data values}
}}
\usage{
HeatExchange
}
\description{
A dataset containing data simulated from a heat exchange simulator
}
\keyword{datasets}
