% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfitpred-class.R
\docType{methods}
\name{plot,bcgpfitpred,ANY-method}
\alias{plot,bcgpfitpred,ANY-method}
\title{Plot a bcgpfitpred object}
\usage{
\S4method{plot}{bcgpfitpred,ANY}(x, ..., decomposition = FALSE,
  print = TRUE)
}
\arguments{
\item{x}{An instance of class \linkS4class{bcgpfitpred}}

\item{...}{optional parameters}

\item{decomposition}{A logical indicating whether to plot the data process
decomposed into global and local processes (TRUE) or the overall process by
itself (FALSE). This parameter is ignored when the data has dimension larger
than 1. Defaults to FALSE}

\item{print}{A logical indicating whether to automatically print the plots.
Defaults to TRUE.}
}
\value{
Either one \code{\link[ggplot2]{ggplot}} object or a list of ggplot
objects that can be further customized using the \pkg{ggplot2} package.
}
\description{
This plots the observed data in a \code{bcgpfitpred} object.
}
\details{
Plotting predictions helps to visualize the predictions at new data locations
and how well the model performs. Plotting is only supported for
one-dimensional data.
}
\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                    d = 1, decomposition = TRUE)

model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling(model, scaled = TRUE, cores = 4, nmcmc = 1000,
                     burnin = 500)
plot(fit)
}
\seealso{
\linkS4class{bcgpfitpred}
}
