% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpsims-class.R
\docType{methods}
\name{plot,bcgpsims,ANY-method}
\alias{plot,bcgpsims,ANY-method}
\title{Plot a bcgpsims object}
\usage{
\S4method{plot}{bcgpsims,ANY}(x, ..., process = c("y", "variance"),
  decomposition = FALSE, print = TRUE)
}
\description{
Plot a bcgpsims object
}
