% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfitpred-class.R
\name{plot_predictions,bcgpfitpred-method}
\alias{plot_predictions,bcgpfitpred-method}
\title{\code{plot_predictions} method for a \code{bcgpfitpred} object}
\usage{
\S4method{plot_predictions}{bcgpfitpred}(
  object,
  prob = 0.95,
  estimate = c("mean", "median"),
  interval_type = c("central", "HPD"),
  overlay_data = TRUE,
  print = TRUE
)
}
\arguments{
\item{object}{An instance of class \linkS4class{bcgpfitpred}}

\item{prob}{A numeric scalar between 0 and 1 specifying the uncertainty
interval width. Defaults to 0.95. \code{NULL} will give no uncertainty
intervals.}

\item{estimate}{A character specifying whether the mean or median should be
used as the point estimate. Defaults to "mean".}

\item{interval_type}{A character specifying whether "central" intervals based
on quantiles or highest posterior density ("HPD") intervals should be
constructed. Defaults to "central". If \code{prob} is \code{NULL} then this
argument is ignored.}

\item{overlay_data}{A logical specifying whether the observed data should be
overlaid. Defaults to TRUE.}

\item{print}{A logical indicating whether to automatically print the plots.
Defaults to TRUE.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further customized
 using the \pkg{ggplot2} package.
}
\description{
This plots the predictions in a \code{bcgpfitpred} object.
}
\details{
Plotting the process \eqn{Y(x)}, helps to visualize the predictions.
Plotting is only supported for one-dimensional data.
}
\examples{
data_sim <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                     d = 2, decomposition = TRUE)

model <- bcgpmodel(x = data_sim@training$x, y = data_sim@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling_with_predictions(model, x_pred = data_sim@test$x,
                                      scaled = FALSE, cores = 4,
                                      n_mcmc = 500, burnin = 200)

plot(fit, prob = 0.90, esimate = "mean", overlay_data = TRUE,
     interval_type = "central")
}
\seealso{
\linkS4class{bcgpfitpred}
}
