% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfit-class.R
\name{posterior_interval,bcgpfit-method}
\alias{posterior_interval,bcgpfit-method}
\title{\code{posterior_interval} method for a \code{bcgpfit} object}
\usage{
\S4method{posterior_interval}{bcgpfit}(object, prob = 0.95, type = c("central", "HPD"), pars)
}
\arguments{
\item{object}{a bcgpfitpred object}

\item{prob}{A numeric scalar between 0 and 1 specifying the uncertainty
interval width. Defaults to 0.95.}

\item{type}{A character specifying whether "central" intervals based on
quantiles or highest posterior density ("HPD") intervals should be
constructed. Defaults to "central".}

\item{pars}{A character vector specifying the parameters for which to return
intervals.}
}
\value{
A matrix with two columns and as many rows as model parameters
(or a subset of parameters specified by the user). For a given value of
\code{prob}, \eqn{p}, the columns correspond to the lower and upper
\eqn{100p}\% interval limits and have the names \code{lower} and
\code{upper}.
}
\description{
This calculates Bayesian uncertainty intervals, commonly called
\emph{credible} intervals
}
\examples{
data_sim <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                     d = 2, decomposition = TRUE)

model <- bcgpmodel(x = data_sim@training$x, y = data_sim@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling(model, scaled = FALSE, cores = 4,
                     n_mcmc = 500, burnin = 200)

posterior_interval(fit, prob = 0.90, type = "central",
                   pars = c("beta0", "w", "rhoG", "rhoL"))

}
