% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfit-class.R
\docType{methods}
\name{posterior_predict,bcgpfit-method}
\alias{posterior_predict,bcgpfit-method}
\title{\code{posterior_predict} method for a \code{bcgpfit} object}
\usage{
\S4method{posterior_predict}{bcgpfit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a bcgpfit object}

\item{newdata}{Optionally, an \emph{nPred x d} numeric matrix of new data
locations at which to predict. If omitted, the training data matrix is used.
If \code{newdata} is provided, it should be provided on the same scale as the
user-provided training data, i.e. do not transform to \eqn{[0, 1]^d}.}
}
\value{
A list with elements \code{x}, an \code{nPred x d} numeric matrix
representing the prediction locations and \code{y}, an \code{iter x nPred}
matrix of simulations from the posterior predictive distribution. Each row of
the matrix is a vector of predictions generated using a single draw of the
model parameters from the posterior distribution.
}
\description{
This makes posterior predictions for either new data or for the training
data.
}
\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                    d = 1, decomposition = TRUE)

model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling(model, scaled = TRUE, cores = 4, nmcmc = 500,
                     burnin = 200, algorithm = "NUTS",
                     control = list(adapt_delta = 0.90))

posterior_predict(fit)
posterior_predict(fit, newdata = matrix(runif(100, -0.5, 1.5), ncol = 1))

}
