% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfit-class.R
\docType{methods}
\name{predict,bcgpfit-method}
\alias{predict,bcgpfit-method}
\title{\code{predict} method for a \code{bcgpfit} object}
\usage{
\S4method{predict}{bcgpfit}(object, newdata = NULL, prob = 0.95, ...)
}
\arguments{
\item{object}{a bcgpfit object}

\item{newdata}{Optionally, an \emph{nPred x d} numeric matrix of new data
locations at which to predict. If omitted, the training data matrix is used.
If \code{newdata} is provided, it should be provided on the same scale as the
user-provided training data, i.e. do not transform to \eqn{[0, 1]^d}.}

\item{prob}{a single number greater than 0 and less than 1 that specifes the
width of the (equal-tailed) posterior prediction interval. For example,
\code{prob = 0.90} specifies that 90\% prediction intervals are desired.}
}
\value{
An instance of S4 class \code{bcgpfitpred}
}
\description{
This function computes Bayesian posterior predictions and prediction
intervals.
}
\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                    d = 2, decomposition = TRUE)

model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling(model, scaled = TRUE, cores = 4, nmcmc = 500,
                     burnin = 200)

fit
print(fit, pars = c("beta0", "w", "rhoG", "rhoL"), digits_summary = 3)
predict(fit)
}
\seealso{
\linkS4class{bcgpmodel} \linkS4class{bcgpfit}
\code{\link[bcgp]{posterior_predict}}
}
