% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescaleXPred}
\alias{rescaleXPred}
\title{Rescale a vector}
\usage{
rescaleXPred(newdata, x)
}
\arguments{
\item{\code{newdata}}{A vector input}

\item{\code{x}}{A vector input that is scaled to}
}
\value{
A vector rescaled to [0, 1]
}
\description{
Rescales a vector to the same scale as a vector that has already
been scaled to [0,1]
}
\details{
Rescales a vector to the same scale as a vector that has already
been scaled to [0,1]
}
\examples{
rescaleXPred(rnorm(10, 0, 100))
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
