% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfitpred-class.R
\name{rmspe,bcgpfitpred-method}
\alias{rmspe,bcgpfitpred-method}
\title{\code{rmspe} method for a \code{bcgpfitpred} object}
\usage{
\S4method{rmspe}{bcgpfitpred}(object, truth, pred_type = c("mean", "median"))
}
\arguments{
\item{object}{a bcgpfitpred object}

\item{truth}{a numeric vector specifying the true values that should be
compared to the \code{bcgp} predictions}

\item{pred_type}{a character specifying the posterior point estimator to use.
It can be either "mean" or "median". Defaults to "mean}
}
\description{
This calculates the root mean square prediction error for the posterior
predictions contained in a \code{bcgpfitpred} object in comparison to a
user-provided vector
}
\examples{
data_sim <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                     d = 2, decomposition = TRUE)

model <- bcgpmodel(x = data_sim@training$x, y = data_sim@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling_with_predictions(model, x_pred = data_sim@test$x,
                                      scaled = FALSE, cores = 4,
                                      n_mcmc = 500, burnin = 200)

rmspe(fit, truth = data_sim@test$y, pred_type = "mean")

}
