% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{scaleX}
\alias{scaleX}
\title{Scale a matrix to the unit hypercube}
\usage{
scaleX(x)
}
\arguments{
\item{\code{x}}{An \code{n x d} matrix}
}
\value{
A matrix rescaled to \deqn{[0, 1]^d} with attributes
\code{"scaled:minimum"} and \code{"scaled:range"}
}
\description{
Scales a matrix to \deqn{[0, 1]^d}
}
\details{
This function scales a matrix to the unit hypercube by subtracting off
the minimum value in each column from each value and dividing by the range of
that column.
}
\examples{
scaleX(matrix(runif(20, 0, 10), nrow = 10, ncol = 2))
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
