% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{scaleXPred}
\alias{scaleXPred}
\title{Scale a matrix to the same scale as the scaled training data matrix}
\usage{
scaleXPred(x)
}
\arguments{
\item{\code{newdata}}{An \code{n x d} matrix}

\item{\code{x}}{An \code{n x d} matrix that has already been scaled to
\eqn{[0, 1]^d}. This matrix needs to have attributes \code{"scaled:minimum"}
and \code{"scaled:range"}}
}
\value{
A matrix rescaled to \deqn{[0, 1]^d} with attributes
\code{"scaled:minimum"} and \code{"scaled:range"}
}
\description{
Scales a matrix to the same scale as the scaled training data
matrix
}
\details{
This function scales a matrix to the same scale as the scaled
training data matrix by subtracting off the minimum value of the training
data in each column from each value and dividing by the range of the training
data in that column.
}
\examples{
scaleXPred(matrix(runif(20, 0, 10), nrow = 50, ncol = 2),
           scaleX(matrix(runif(20, 0, 5), nrow = 10, ncol = 2)))
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
