% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFuncs.R
\name{simulate_from_model}
\alias{simulate_from_model}
\title{Simulate from the model}
\usage{
simulate_from_model(composite = TRUE, stationary = FALSE,
  noise = FALSE, d = 1, n = 15 * d, nPred = 100 * d,
  parameters = createParameterList(composite, stationary, noise, d))
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process
and a local process, \code{FALSE} for non-composite. Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}

\item{n}{An integer giving the desired number of training data locations.}

\item{nPred}{An integer giving the desired number of test data locations.}

\item{parameters}{A list containing desired parameter values. If missing,
then parameter values will be drawn at random. A call to
\code{createParameterList()} will assist in the correct creation of this
list.}
}
\value{
A list with elements \code{x} (an \emph{n x d} matrix), \code{y} (a
vector of length \emph{n}), \code{xPred} (an \emph{nPred x d} matrix), and
\code{yPred} (a vector of length \emph{nPred}) corresponding to training
locations, training responses, test locations, and test responses,
respectively, and element \code{parameters}, which contains the parameters
used to simulate the data.
}
\description{
This function simulates data from the specified (stationary, composite,
noise) Gaussian process model.
}
\details{
\code{simulate_from_model} returns a list with elements \code{x}, \code{y},
\code{xPred}, \code{yPred}, corresponding to training locations, training
responses, test locations, and test responses, respectively, and element
\code{parameters}, which contains the parameters used to simulate the data.
}
\examples{
simulate_from_model(composite = TRUE, stationary = FALSE, noise = FALSE)
}
\seealso{
\code{\link{createParameterList}}
}
