% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfitpred-class.R
\name{summary_predictions,bcgpfitpred-method}
\alias{summary_predictions,bcgpfitpred-method}
\title{\code{summary_predictions} method for a \code{bcgpfitpred} object}
\usage{
\S4method{summary_predictions}{bcgpfitpred}(
  object,
  pars = c("y", "V"),
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)
}
\arguments{
\item{object}{a bcgpfitpred object}

\item{pars}{a character vector specifying the parameters to summarize. One or
both of "y" and "V" may be specified to summarize the posterior predictions
of the process or the variance process, respectively. If the process is
stationary, then thie argument will be ignored.}

\item{quantiles}{a numeric vector specifying the desired quantiles}
}
\description{
This gives a summary of the posterior predictions contained in a
\code{bcgpfitpred} object
}
\examples{
data_sim <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                     d = 2, decomposition = TRUE)

model <- bcgpmodel(x = data_sim@training$x, y = data_sim@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling_with_predictions(model, x_pred = data_sim@test$x,
                                      scaled = FALSE, cores = 4,
                                      n_mcmc = 500, burnin = 200)

fit
print(fit, pars = c("beta0", "w", "rhoG", "rhoL"), digits_summary = 3)
print_predictions(fit, pars = c("y", "V"), digits_summary = 3)
summary(fit)
summary_predictions(fit)

}
