% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{unscaleY}
\alias{unscaleY}
\title{Unscale a response vector}
\usage{
unscaleY(y, yTrain)
}
\arguments{
\item{\code{y}}{A numeric vector of length \code{nPred}}

\item{\code{yTrain}}{A numeric vector of length \code{nPred} that has already
been scaled. This vector needs to have attributes \code{"scaled:center"}
and \code{"scaled:scale"}}
}
\value{
A vector rescaled to it's natural scale with attributes
\code{"scaled:minimum"} and \code{"scaled:range"}
}
\description{
Unscales a response vector to the same scale as the scaled
training data response vector
}
\details{
This function scales a vector to the same scale as the scaled
training data response vector by multiplying the response by the standard
deviation of the training data and adding the mean of the training data
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
