% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpmodel-class.R
\docType{methods}
\name{update_inits,bcgpmodel-method}
\alias{update_inits,bcgpmodel-method}
\title{Update the starting values for MCMC chains}
\usage{
\S4method{update_inits}{bcgpmodel}(object)
}
\arguments{
\item{object}{An instance of S4 class \code{bcgpmodel}.}
}
\value{
An instance of S4 class \code{bcgpmodel} containing initial values
randomly from the prior distributions for all the parameters
}
\description{
\code{update_inits} returns an instance of S4 class \code{bcgpmodel} after
updating the starting values for each Markov chain
}
\details{
This returns an instance of S4 class \code{bcgpmodel}. It is generally meant
to update the starting values for the Markv chains after the user has changed
the prior distributions in an instance of class \code{bcgpmodel}. The new
initial values are simulated from the prior distribution, ensure they are
within the support of the posterior distribution.
}
\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE,
                   scaled = TRUE, chains = 4L)
model@priors$w$lower <- 0.75
model <- update_inits(model)
}
\seealso{
\linkS4class{bcgpmodel}

Other preprocessing functions: \code{\link{bcgpmodel}},
  \code{\link{create_parameter_list}}
}
\concept{preprocessing functions}
