% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{WingWeight}
\alias{WingWeight}
\title{Data from the Wing Weight example}
\format{A list with components for training data and test data:
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{x_test}{test data locations}
  \item{y_test}{test data values}
}}
\usage{
WingWeight
}
\description{
A dataset containing data for the wing weight of a light aircraft
}
\keyword{datasets}
