% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WW}
\alias{WW}
\title{Wing Weight Data}
\format{A list with elements \code{x, y, xTest, yTest}
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{xTest}{test data locations}
  \item{yTest}{test data values}
}}
\source{
TODO: add source
}
\usage{
WW
}
\description{
A dataset containing the training data and test data from
the wing weight example used in the paper
}
\keyword{datasets}
