% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpmodel-class.R
\name{bcgpmodel}
\alias{bcgpmodel}
\title{Create an object of class bcgpmodel}
\usage{
bcgpmodel(x, y, stationary = FALSE, composite = TRUE, noise = FALSE,
  scaled = TRUE, chains = 4L)
}
\arguments{
\item{x}{An \emph{n x d} numeric matrix, where \emph{d} is the dimension of
the data, and \emph{n} is the number of training data observations.}

\item{y}{A numeric vector of length \emph{n}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{composite}{A logical, \code{TRUE} for a composite of a global process,
a local process, and an error process, \code{FALSE} for non-composite.
Defaults to \code{TRUE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{scaled}{A logical indicating whether the user would like the data to be
scaled before fitting.}

\item{chains}{A positive integer specifying the number of Markov chains}
}
\value{
An instance of S4 class \code{bcgpmodel} containing the default
values for all the prior parameters, information about the process, and
information about the distributions.
}
\description{
\code{bcgpmodel} returns an instance of S4 class \code{bcgpmodel}
}
\details{
This creates an instance of S4 class \code{bcgpmodel} that contains
user-provided data, user-provided information about the type of model desired
to be fit (composite/non-composite, stationary/non-stationary,
deterministic/noisy), default values for the hyperparameters for the prior
distributions, information about the prior distributions, randomly-generated
starting values for the Markov chains, user-provided information about
whether the user desires the data to be scaled before it is fit, and user-
provided information on the number of chains desired to be sampled.
}
\examples{
simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE)
bcgpmodel(x = simData@training$x, y = simData@training$y,
          composite = TRUE, stationary = FALSE, noise = FALSE,
          scaled = TRUE, chains = 4L)
}
\seealso{
\linkS4class{bcgpmodel} \code{\link{bcgpfit}}
\code{\link{bcgp}}

Other preprocessing functions: \code{\link{create_parameter_list}},
  \code{\link{update_inits,bcgpmodel-method}}
}
\concept{preprocessing functions}
