% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BJX}
\alias{BJX}
\title{BJX Data}
\format{A list with elements \code{x, y, xTest, yTest}
\describe{
  \item{x}{training data locations}
  \item{y}{training data values}
  \item{xTest}{test data locations}
  \item{yTest}{test data values}
}}
\source{
TODO: add source
}
\usage{
BJX
}
\description{
A dataset containing the training data and test data from
the BJX function \deqn{y(x) = sin\left(30\left( x - 0.9\right)^4 \right)
cos\left(2\left( x - 0.9\right)\right) + \frac{x - 0.9}{2}, \;\;\;\;
x \in [0, 1]} used in the paper
}
\keyword{datasets}
