% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFuncs.R
\name{create_priors}
\alias{create_priors}
\title{Create an object of class bcgppriors}
\usage{
create_priors(composite = TRUE, stationary = FALSE, noise = FALSE,
  d = 1L)
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process,
a local process, and an error process, \code{FALSE} for non-composite.
Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}
}
\value{
An instance of S4 class \code{bcgppriors} containing the default
values for all the prior parameters, information about the process, and
information about the distributions.
}
\description{
\code{create_priors} returns an instance of S4 class \code{bcgppriors}
}
\details{
This creates an instance of S4 class \code{bcgppriors} that contains default
values for the priors, information about the process, and information about
the distributions. The user can change the values as they like prior to
fitting any data with \code{\link{bcgp}}. \code{bcgppriors()} can also be
called to create a \code{bcgppriors} object.
}
\examples{
create_priors(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1)
create_priors(composite = FALSE, stationary = TRUE, noise = TRUE, d = 3)
}
\seealso{
\linkS4class{bcgppriors} \code{\link{bcgppriors}}
\code{\link{bcgp}}
}
