% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFuncs.R
\name{createParameterList}
\alias{createParameterList}
\title{Create a list with random parameter values.}
\usage{
createParameterList(composite = TRUE, stationary = FALSE,
  noise = FALSE, d = 1)
}
\arguments{
\item{composite}{A logical, \code{TRUE} for a composite of a global process
and a local process, \code{FALSE} for non-composite. Defaults to \code{TRUE}.}

\item{stationary}{A logical, \code{FALSE} for a non-stationary process,
\code{TRUE} for a stationary process. If \code{FALSE}, the variance for the
process is \eqn{\sigma^2(x)}, and if \code{TRUE}, the variance is
\eqn{\sigma^2}. Defaults to \code{FALSE}.}

\item{noise}{If the data should be noise-free (such as from a deterministic
computer model), then \code{noise} should be \code{FALSE}. Otherwise, it
should be \code{TRUE}. Defaults to \code{FALSE}}

\item{d}{An integer giving the dimension of the data.}
}
\value{
A list of randomly-generated parameter values corresponding to the
BCGP model described by \code{composite}, \code{stationary}, and
\code{noise}.
}
\description{
\code{createParameterList} returns a list that contains randomly generated
parameter values.
}
\details{
This creates a list that contains randomly generated parameter values for a
model corresponding to inputs \code{composite}, \code{stationary},
\code{noise}, \code{d}. This function is meant to set up the parameters
before a call to \code{simulate_from_model}. The user can modify the list of
parameter values if desired. If a non-stationary model is desired, the
parameters that define the variance process,
(\eqn{[\mu_V,\rho_V,\sigma^2_V]^\top}), will be returned and can then be
modified, but the variance process itself will not be returned and cannot be
specified.
}
\examples{
createParameterList(composite = FALSE, stationary = TRUE, noise = FALSE, d = 1)
createParameterList(composite = TRUE, stationary = FALSE, noise = FALSE, d = 1)
}
\seealso{
\code{\link{simulate_from_model}}

Other preprocessing functions: \code{\link{createInits}},
  \code{\link{createPriors}}
}
\concept{preprocessing functions}
