% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpsims-class.R
\docType{methods}
\name{plot,bcgpsims,ANY-method}
\alias{plot,bcgpsims,ANY-method}
\title{Plot a bcgpsims object}
\usage{
\S4method{plot}{bcgpsims,ANY}(x, ..., process = c("y", "variance"),
  decomposition = FALSE, print = TRUE)
}
\arguments{
\item{x}{An instance of class \linkS4class{bcgpsims}}

\item{...}{optional parameters
\describe{
  \item{\code{raster}}{A logical indicating whether the 2-dimensional process
  should be plotted in points or smoothed out as in
  \code{\link[ggplot2]{geom_raster}}. If the \code{bcgpsims} object was
  created with randomly generated test data (i.e., not in a grid), this
  argument is ignored, and the plot will be points and not smoothed out.
  Only useful for 2-D data.}
}}

\item{process}{Which process to plot: the data process and/or the variance
process}

\item{decomposition}{A logical indicating whether to plot the data process
decomposed into global and local processes (TRUE) or the overall process by
itself (FALSE). This parameter is ignored when the data has dimension larger
than 1. Defaults to FALSE}

\item{print}{A logical indicating whether to automatically print the plots.
Defaults to TRUE.}
}
\value{
Either one \code{\link[ggplot2]{ggplot}} object or a list of ggplot
objects that can be further customized using the \pkg{ggplot2} package.
}
\description{
This plots the simulated data in a \code{bcgpsims} object.
}
\details{
Plotting simulated data helps to understand what these types of models look
like and understand the interaction between the variance process
\eqn{\sigma^2(x)} and the data process \eqn{Y(x)} for non-stationary models.
Plotting is only supported for one and two-dimensional data.
}
\examples{
nsComp <- simulate_from_model(composite = TRUE, stationary = FALSE, d = 1,
                              decomposition = TRUE)
z <- plot(nsComp, print = FALSE, decomposition = TRUE)

sNonComp <- simulate_from_model(composite = FALSE, stationary = TRUE, d = 2,
                                gridTest = TRUE, gridTestSize = 10)
plot(sNonComp, process = "y", raster = TRUE)
plot(sNonComp, process = "y", raster = FALSE)
}
\seealso{
\linkS4class{bcgpsims}
}
