% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfitpred-class.R
\docType{methods}
\name{rmspe,bcgpfitpred-method}
\alias{rmspe,bcgpfitpred-method}
\title{Calculate the Square Root of the Mean Squared Prediction Error}
\usage{
\S4method{rmspe}{bcgpfitpred}(object, truth)
}
\arguments{
\item{truth}{A numeric vector of length \code{n_{pred}}}

\item{x}{An instance of class \linkS4class{bcgpfitpred}}
}
\value{
A single number representing the RMSPE
}
\description{
This calculates the square root of the mean squared prediction errer (RMSPE)
for the predictions in a \code{bcgpfitpred} object.
}
\details{
In cases where we have test data, one measure of the accuracy of the
predictions is the RMSPE. It is calculated by
\deqn{RMSPE = \sqrt{\frac{1}{n_{pred}}\sum \left( \hat{y} - y  \right)^2}}
}
\seealso{
\linkS4class{bcgpfitpred}
}
