% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfit-class.R
\docType{methods}
\name{summary,bcgpfit-method}
\alias{summary,bcgpfit-method}
\title{\code{summary} method for a \code{bcgpfit} object}
\usage{
\S4method{summary}{bcgpfit}(object, pars, quantiles = c(0.025, 0.25, 0.5,
  0.75, 0.975), ...)
}
\arguments{
\item{object}{a bcgpfit object}

\item{pars}{a character vector specifying the parameters to summarize}

\item{quantiles}{a numeric vector specifying the desired quantiles for each
parameter}
}
\description{
This gives a summary of the posterior draws contained in a \code{bcgpfit}
object
}
\examples{


simData <- bcgpsims(composite = TRUE, stationary = FALSE, noise = FALSE,
                    d = 2, decomposition = TRUE)

model <- bcgpmodel(x = simData@training$x, y = simData@training$y,
                   composite = TRUE, stationary = FALSE, noise = FALSE)
fit <- bcgp_sampling(model, scaled = TRUE, cores = 4, nmcmc = 500,
                     burnin = 200)

fit
print(fit, pars = c("beta0", "w", "rhoG", "rhoL"), digits_summary = 3)
summary(fit)

}
